# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Harency ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/send_otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send Otp",
                "description": "Send otp to verify user",
                "operationId": "authSendOtp",
                "requestBody": {
                    "description": "Pass user phone number and type if its 0 or 1,0 for verify user and 1 for forgot password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "type"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "format": "phone_number",
                                        "example": "9786543210"
                                    },
                                    "type": {
                                        "type": "string",
                                        "format": "type",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong Phone number response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone number. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/reset_password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset Password",
                "description": "Reset user password",
                "operationId": "authResetpassword",
                "requestBody": {
                    "description": "Pass user password and otp",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "otp"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "format": "otp",
                                        "example": "046677"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong otp number. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/upload_image": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Upload Image",
                "description": "Upload Image",
                "operationId": "authUploadImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Image to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong format response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong Format. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/create_profile": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create Profile",
                "description": "Create profile for user",
                "operationId": "authCreateProfile",
                "requestBody": {
                    "description": "Pass user name and email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "john"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "abc@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong invalid email or name. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}